+++ src/main/java/user/DaoFactoryForTest.java
@@ -1,3 +1,5 @@
+package user;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.datasource.SimpleDriverDataSource;
@@ -5,6 +7,7 @@ import user.connection.ConnectionMaker;
 import user.connection.TestConnectionMaker;
 import user.dao.JdbcContext;
 import user.dao.UserDaoJdbc;
+import user.service.UserService;
 
 import javax.sql.DataSource;
 import java.sql.Driver;
@@ -12,6 +15,11 @@ import java.sql.SQLException;
 
 @Configuration
 public class DaoFactoryForTest {
+    @Bean
+    public UserService userService() throws SQLException {
+        return new UserService(userDao());
+    }
+
     @Bean
     public UserDaoJdbc userDao() throws SQLException {
         return new UserDaoJdbc(dataSource());
