+++ src/main/java/user/service/UserService.java
@@ -1,0 +1,11 @@
+package user.service;
+
+import user.dao.UserDao;
+
+public class UserService {
+    UserDao userDao;
+
+    public UserService(UserDao userDao) {
+        this.userDao = userDao;
+    }
+}
