+++ src/test/java/user/service/UserServiceTest.java
@@ -1,0 +1,22 @@
+package user.service;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import user.DaoFactoryForTest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(classes = DaoFactoryForTest.class)
+public class UserServiceTest {
+    @Autowired
+    UserService userService;
+
+    @Test
+    public void bean() {
+        assertThat(userService).isNotNull();
+    }
+}
